@@ -6,33 +6,32 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.8</version>
+        <version>3.1.2</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <groupId>com.codingapi.springboot</groupId>
     <artifactId>springboot-parent</artifactId>
-    <version>1.5.11</version>
+    <version>2.1.5</version>
 
     <url>https://github.com/codingapi/springboot-framewrok</url>
     <name>springboot-parent</name>
     <description>springboot-parent project for Spring Boot</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
 
         <!-- maven properties-->
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm:ss</maven.build.timestamp.format>
-        <maven.compile.source>1.8</maven.compile.source>
-        <maven.compile.target>1.8</maven.compile.target>
+        <maven.compile.source>17</maven.compile.source>
+        <maven.compile.target>17</maven.compile.target>
         <maven.source.plugin>3.0.1</maven.source.plugin>
-        <maven-compiler-plugin.version>3.6.0</maven-compiler-plugin.version>
-        <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
+        <maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
         <maven.javadoc.plugin>2.10.3</maven.javadoc.plugin>
         <nexus.staging.maven.plugin>1.6.3</nexus.staging.maven.plugin>
-
+        <maven.gpg.plugin>1.6</maven.gpg.plugin>
         <codingapi.framework.version>${project.version}</codingapi.framework.version>
-        <fastjson.version>2.0.23</fastjson.version>
+        <fastjson.version>2.0.28</fastjson.version>
         <jsonwebtoken.jjwt.version>0.11.5</jsonwebtoken.jjwt.version>
         <commons-io.version>2.11.0</commons-io.version>
         <commons-dbutils.version>1.7</commons-dbutils.version>
