@@ -5,14 +5,14 @@
     <parent>
         <artifactId>springboot-parent</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>1.5.11</version>
+        <version>2.1.5</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>springboot-starter-data-fast</artifactId>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
 
 
