@@ -6,7 +6,7 @@
     <parent>
         <artifactId>springboot-parent</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>1.5.11</version>
+        <version>2.1.5</version>
     </parent>
 
     <artifactId>springboot-starter-security-jwt</artifactId>
@@ -15,7 +15,7 @@
     <description>springboot-starter-security-jwt project for Spring Boot</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
 
     <dependencies>
