@@ -3,7 +3,6 @@ package com.codingapi.springboot.security;
 import com.codingapi.springboot.security.configurer.HttpSecurityConfigurer;
 import com.codingapi.springboot.security.controller.VersionController;
 import com.codingapi.springboot.security.filter.*;
-import com.codingapi.springboot.security.handler.ServletExceptionHandler;
 import com.codingapi.springboot.security.jwt.Jwt;
 import com.codingapi.springboot.security.properties.SecurityJwtProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -13,8 +12,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -22,12 +21,11 @@ import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.servlet.HandlerExceptionResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-@EnableMethodSecurity
+@EnableWebSecurity
 public class AutoConfiguration {
 
     @Bean
@@ -57,36 +55,38 @@ public class AutoConfiguration {
 
 
     @Bean
-    public HandlerExceptionResolver servletExceptionHandler() {
-        return new ServletExceptionHandler();
-    }
-
-    @Bean
     @ConditionalOnMissingBean
     public SecurityLoginHandler securityLoginHandler(){
         return (request, response, handler) -> {
-
         };
     }
 
+
     @Bean
     @ConditionalOnMissingBean
-    public SecurityFilterChain filterChain(HttpSecurity http, Jwt jwt,SecurityLoginHandler loginHandler,
+    public SecurityFilterChain filterChain(HttpSecurity security, Jwt jwt,SecurityLoginHandler loginHandler,
                                            SecurityJwtProperties properties) throws Exception {
+        //disable basic auth
+        security.httpBasic().disable();
+
         //before add addCorsMappings to enable cors.
-        http.cors();
+        security.cors();
         if(properties.isDisableCsrf() ){
-            http.csrf().disable();
+            security.csrf().disable();
         }
-        http.apply(new HttpSecurityConfigurer(jwt,loginHandler,properties));
-        http
+        security.apply(new HttpSecurityConfigurer(jwt,loginHandler,properties));
+        security
                 .exceptionHandling()
                 .authenticationEntryPoint(new MyUnAuthenticationEntryPoint())
                 .accessDeniedHandler(new MyAccessDeniedHandler())
                 .and()
-                .authorizeRequests()
-                .antMatchers(properties.getAuthenticatedUrls()).authenticated()
-                .and()
+                .authorizeHttpRequests(
+                        registry -> {
+                            registry.requestMatchers(properties.getIgnoreUrls()).permitAll()
+                                    .requestMatchers(properties.getAuthenticatedUrls()).authenticated()
+                                    .anyRequest().permitAll();
+                        }
+                )
                 //default login url :/login
                 .formLogin()
                 .loginProcessingUrl(properties.getLoginProcessingUrl())
@@ -96,10 +96,9 @@ public class AutoConfiguration {
                 .logout()
                 .logoutUrl(properties.getLogoutUrl())
                 .addLogoutHandler(new MyLogoutHandler())
-                .logoutSuccessHandler(new MyLogoutSuccessHandler())
-                .permitAll();
+                .logoutSuccessHandler(new MyLogoutSuccessHandler());
 
-        return http.build();
+        return security.build();
     }
 
 
