@@ -2,8 +2,8 @@ package com.codingapi.springboot.security.filter;
 
 import com.codingapi.springboot.security.dto.request.LoginRequest;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 public interface SecurityLoginHandler {
 
