@@ -1,7 +1,7 @@
 package com.codingapi.springboot.generator;
 
 import com.codingapi.springboot.generator.dao.IdKeyDao;
-import com.codingapi.springboot.generator.properties.GeneratorProperties;
+import com.codingapi.springboot.generator.properties.GeneratorJdbcProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -11,15 +11,15 @@ import org.springframework.context.annotation.Configuration;
 public class AutoConfiguration {
 
     @Bean
-    @ConfigurationProperties(prefix = "codingapi.id.generator")
-    public GeneratorProperties generatorProperties() {
-        return new GeneratorProperties();
+    @ConfigurationProperties(prefix = "codingapi.id.jdbc.generator")
+    public GeneratorJdbcProperties generatorJdbcProperties() {
+        return new GeneratorJdbcProperties();
     }
 
     @Bean(initMethod = "init")
     @ConditionalOnMissingBean
-    public IdKeyDao idKeyDao(GeneratorProperties generatorProperties) {
-        IdKeyDao keyDao = new IdKeyDao(generatorProperties.getJdbcUrl());
+    public IdKeyDao idKeyDao(GeneratorJdbcProperties generatorJdbcProperties) {
+        IdKeyDao keyDao = new IdKeyDao(generatorJdbcProperties.openDataSource());
         IdGeneratorContext.getInstance().init(keyDao);
         return keyDao;
     }
