@@ -14,8 +14,8 @@ public class RestTemplateContext {
     private final RestTemplate restTemplate;
 
     private RestTemplateContext(){
-        HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(
-                TrustAnyHttpClientFactory.createTrustAnyHttpClient());
+        HttpComponentsClientHttpRequestFactory requestFactory
+                = new HttpComponentsClientHttpRequestFactory(TrustAnyHttpClientFactory.createTrustAnyHttpClient());
         this.restTemplate = restTemplate(requestFactory);
     }
 
