@@ -1,10 +1,10 @@
 package com.codingapi.springboot.framework.crypto;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Base64Utils;
 
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
+import java.util.Base64;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -12,7 +12,7 @@ class HmacSHA256Test {
 
     @Test
     void sha256() throws NoSuchAlgorithmException, InvalidKeyException {
-       String data = Base64Utils.encodeToString(HmacSHA256.sha256("123456".getBytes(),"123456".getBytes()));
+       String data = Base64.getEncoder().encodeToString(HmacSHA256.sha256("123456".getBytes(),"123456".getBytes()));
        assertEquals("uK0Io6VH41gpuCG3U3AwHdjEsGvdd3H5tUGnWRQGhxg=",data);
     }
 }
